fix : S3 bucket file is moved to lib
