+++ actions/S3/uploadFileToS3.ts
@@ -2,7 +2,7 @@
 
 import { PutObjectCommand, GetObjectCommand } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
-import s3 from "./bucket";
+import s3 from "../../lib/bucket";
 import crypto from "crypto";
 const generateFileName = (bytes = 32) =>
   crypto.randomBytes(bytes).toString("hex");
